# backend/api/gen/go/admin/service/v1/i_balance_flow_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_balance_flow.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-ledger/api/gen/go/ledger/service/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationBalanceFlowServiceConfirm = "/admin.service.v1.BalanceFlowService/Confirm"
const OperationBalanceFlowServiceCreate = "/admin.service.v1.BalanceFlowService/Create"
const OperationBalanceFlowServiceDelete = "/admin.service.v1.BalanceFlowService/Delete"
const OperationBalanceFlowServiceGet = "/admin.service.v1.BalanceFlowService/Get"
const OperationBalanceFlowServiceList = "/admin.service.v1.BalanceFlowService/List"
const OperationBalanceFlowServiceStatistics = "/admin.service.v1.BalanceFlowService/Statistics"
const OperationBalanceFlowServiceUpdate = "/admin.service.v1.BalanceFlowService/Update"

type BalanceFlowServiceHTTPServer interface {
	Confirm(context.Context, *v11.ConfirmBalanceFlowRequest) (*v11.BalanceFlow, error)
	Create(context.Context, *v11.CreateBalanceFlowRequest) (*v11.BalanceFlow, error)
	Delete(context.Context, *v11.DeleteBalanceFlowRequest) (*emptypb.Empty, error)
	Get(context.Context, *v11.GetBalanceFlowRequest) (*v11.BalanceFlow, error)
	List(context.Context, *v1.PagingRequest) (*v11.ListBalanceFlowResponse, error)
	Statistics(context.Context, *v11.StatisticsRequest) (*v11.StatisticsResponse, error)
	Update(context.Context, *v11.UpdateBalanceFlowRequest) (*v11.BalanceFlow, error)
}

func RegisterBalanceFlowServiceHTTPServer(s *http.Server, srv BalanceFlowServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/balance-flows", _BalanceFlowService_List3_HTTP_Handler(srv))
	r.GET("/admin/v1/balance-flows/{id}", _BalanceFlowService_Get3_HTTP_Handler(srv))
	r.POST("/admin/v1/balance-flows", _BalanceFlowService_Create2_HTTP_Handler(srv))
	r.PUT("/admin/v1/balance-flows/{id}", _BalanceFlowService_Update2_HTTP_Handler(srv))
	r.DELETE("/admin/v1/balance-flows/{id}", _BalanceFlowService_Delete2_HTTP_Handler(srv))
	r.PATCH("/admin/v1/balance-flows/{id}/confirm", _BalanceFlowService_Confirm0_HTTP_Handler(srv))
	r.GET("/admin/v1/balance-flow-statistics", _BalanceFlowService_Statistics1_HTTP_Handler(srv))
}

func _BalanceFlowService_List3_HTTP_Handler(srv BalanceFlowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBalanceFlowServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListBalanceFlowResponse)
		return ctx.Result(200, reply)
	}
}

func _BalanceFlowService_Get3_HTTP_Handler(srv BalanceFlowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetBalanceFlowRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBalanceFlowServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetBalanceFlowRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.BalanceFlow)
		return ctx.Result(200, reply)
	}
}

func _BalanceFlowService_Create2_HTTP_Handler(srv BalanceFlowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreateBalanceFlowRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBalanceFlowServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*v11.CreateBalanceFlowRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.BalanceFlow)
		return ctx.Result(200, reply)
	}
}

func _BalanceFlowService_Update2_HTTP_Handler(srv BalanceFlowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdateBalanceFlowRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBalanceFlowServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdateBalanceFlowRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.BalanceFlow)
		return ctx.Result(200, reply)
	}
}

func _BalanceFlowService_Delete2_HTTP_Handler(srv BalanceFlowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeleteBalanceFlowRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBalanceFlowServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeleteBalanceFlowRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _BalanceFlowService_Confirm0_HTTP_Handler(srv BalanceFlowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.ConfirmBalanceFlowRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBalanceFlowServiceConfirm)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Confirm(ctx, req.(*v11.ConfirmBalanceFlowRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.BalanceFlow)
		return ctx.Result(200, reply)
	}
}

func _BalanceFlowService_Statistics1_HTTP_Handler(srv BalanceFlowServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.StatisticsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBalanceFlowServiceStatistics)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Statistics(ctx, req.(*v11.StatisticsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.StatisticsResponse)
		return ctx.Result(200, reply)
	}
}

type BalanceFlowServiceHTTPClient interface {
	Confirm(ctx context.Context, req *v11.ConfirmBalanceFlowRequest, opts ...http.CallOption) (rsp *v11.BalanceFlow, err error)
	Create(ctx context.Context, req *v11.CreateBalanceFlowRequest, opts ...http.CallOption) (rsp *v11.BalanceFlow, err error)
	Delete(ctx context.Context, req *v11.DeleteBalanceFlowRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	Get(ctx context.Context, req *v11.GetBalanceFlowRequest, opts ...http.CallOption) (rsp *v11.BalanceFlow, err error)
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListBalanceFlowResponse, err error)
	Statistics(ctx context.Context, req *v11.StatisticsRequest, opts ...http.CallOption) (rsp *v11.StatisticsResponse, err error)
	Update(ctx context.Context, req *v11.UpdateBalanceFlowRequest, opts ...http.CallOption) (rsp *v11.BalanceFlow, err error)
}

type BalanceFlowServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewBalanceFlowServiceHTTPClient(client *http.Client) BalanceFlowServiceHTTPClient {
	return &BalanceFlowServiceHTTPClientImpl{client}
}

func (c *BalanceFlowServiceHTTPClientImpl) Confirm(ctx context.Context, in *v11.ConfirmBalanceFlowRequest, opts ...http.CallOption) (*v11.BalanceFlow, error) {
	var out v11.BalanceFlow
	pattern := "/admin/v1/balance-flows/{id}/confirm"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBalanceFlowServiceConfirm))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BalanceFlowServiceHTTPClientImpl) Create(ctx context.Context, in *v11.CreateBalanceFlowRequest, opts ...http.CallOption) (*v11.BalanceFlow, error) {
	var out v11.BalanceFlow
	pattern := "/admin/v1/balance-flows"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBalanceFlowServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BalanceFlowServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeleteBalanceFlowRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/balance-flows/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBalanceFlowServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BalanceFlowServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetBalanceFlowRequest, opts ...http.CallOption) (*v11.BalanceFlow, error) {
	var out v11.BalanceFlow
	pattern := "/admin/v1/balance-flows/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBalanceFlowServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BalanceFlowServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListBalanceFlowResponse, error) {
	var out v11.ListBalanceFlowResponse
	pattern := "/admin/v1/balance-flows"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBalanceFlowServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BalanceFlowServiceHTTPClientImpl) Statistics(ctx context.Context, in *v11.StatisticsRequest, opts ...http.CallOption) (*v11.StatisticsResponse, error) {
	var out v11.StatisticsResponse
	pattern := "/admin/v1/balance-flow-statistics"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBalanceFlowServiceStatistics))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BalanceFlowServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdateBalanceFlowRequest, opts ...http.CallOption) (*v11.BalanceFlow, error) {
	var out v11.BalanceFlow
	pattern := "/admin/v1/balance-flows/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBalanceFlowServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}