# backend/app/app/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Content Hub App API
    description: GoWind Content Hub App API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-ledger
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-ledger/blob/master/LICENSE
    version: "1.0"
paths:
...
    /app/v1/balance-flows:
        get:
            tags:
                - BalanceFlowService
            operationId: BalanceFlowService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.ListBalanceFlowResponse'
        post:
            tags:
                - BalanceFlowService
            operationId: BalanceFlowService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ledger.service.v1.CreateBalanceFlowRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.BalanceFlow'
    /app/v1/balance-flows/{id}:
        get:
            tags:
                - BalanceFlowService
            operationId: BalanceFlowService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.BalanceFlow'
        put:
            tags:
                - BalanceFlowService
            operationId: BalanceFlowService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ledger.service.v1.UpdateBalanceFlowRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.BalanceFlow'
        delete:
            tags:
                - BalanceFlowService
            operationId: BalanceFlowService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
    /app/v1/reports/balance:
        get:
            tags:
                - ReportService
            operationId: ReportService_Balance
            parameters:
                - name: bookId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.BalanceReportResponse'
...
components:
    schemas:
...
        ledger.service.v1.BalanceFlow:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                tenantId:
                    type: integer
                    format: uint32
                bookId:
                    type: integer
                    format: uint32
                type:
                    enum:
                        - FLOW_TYPE_UNSPECIFIED
                        - FLOW_TYPE_EXPENSE
                        - FLOW_TYPE_INCOME
                        - FLOW_TYPE_TRANSFER
                        - FLOW_TYPE_ADJUST
                    type: string
                    description: 流水类型
                    format: enum
                amount:
                    type: string
                    description: 金额
                convertedAmount:
                    type: string
                    description: 换算后的金额
                accountId:
                    type: integer
                    description: 来源账户ID
                    format: uint32
                toAccountId:
                    type: integer
                    description: 目标账户ID（转账）
                    format: uint32
                payeeId:
                    type: integer
                    description: 收款人/付款人ID
                    format: uint32
                creatorId:
                    type: integer
                    format: uint32
                createTime:
                    type: string
                    description: 交易时间（epoch 毫秒）
                title:
                    type: string
                notes:
                    type: string
                confirm:
                    type: boolean
                    description: 是否已确认（余额已变更）
                include:
                    type: boolean
                    description: 是否包含在统计中
                insertAt:
                    type: string
                categories:
                    type: array
                    items:
                        $ref: '#/components/schemas/ledger.service.v1.CategoryRelation'
                    description: 关联的类别拆分
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/ledger.service.v1.TagRelation'
                    description: 关联的标签拆分
                createdBy:
                    type: integer
                    format: uint32
                updatedBy:
                    type: integer
                    format: uint32
                deletedBy:
                    type: integer
                    format: uint32
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                deletedAt:
                    type: string
                    format: date-time
            description: 余额流水
        ledger.service.v1.BalanceReportResponse:
            type: object
            properties:
                assets:
                    type: array
                    items:
                        $ref: '#/components/schemas/ledger.service.v1.ChartPoint'
                    description: 资产明细
                debts:
                    type: array
                    items:
                        $ref: '#/components/schemas/ledger.service.v1.ChartPoint'
                    description: 负债明细
                netWorth:
                    type: string
                    description: 净资（资产-负债）
            description: 资产负债报表回应
...
        ledger.service.v1.CategoryRelation:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                categoryId:
                    type: integer
                    description: 类别ID
                    format: uint32
                amount:
                    type: string
                    description: 金额
                convertedAmount:
                    type: string
                    description: 换算后的金额
            description: 类别关系（一条流水可拆分到多个类别）
...
        ledger.service.v1.ChartPoint:
            type: object
            properties:
                x:
                    type: string
                    description: 标签（分类/标签/收款人名称）
                y:
                    type: string
                    description: 金额
                percent:
                    type: string
                    description: 占比(%)
            description: 图表数据点
...
        ledger.service.v1.CreateBalanceFlowRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/ledger.service.v1.BalanceFlow'
            description: 请求 - 创建流水
...
        ledger.service.v1.ListBalanceFlowResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/ledger.service.v1.BalanceFlow'
                total:
                    type: string
            description: 回应 - 流水列表
...
        ledger.service.v1.TagRelation:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                tagId:
                    type: integer
                    description: 标签ID
                    format: uint32
                amount:
                    type: string
                convertedAmount:
                    type: string
            description: 标签关系
...
        ledger.service.v1.UpdateBalanceFlowRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/ledger.service.v1.BalanceFlow'
                updateMask:
                    type: string
                    format: field-mask
                allowMissing:
                    type: boolean
            description: 请求 - 更新流水
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...